Native style: factor out native style to a separate plugin
This lets us add default QML implementations of the style controls. It's
expected that those files will normally be the same from platform
to platform, since the margins and such are dictated by QStyle. At the
same time, each individual style can still choose to not use the default
QML files, or use it, but override e.g just the background etc.

Change-Id: I13ae2fe42b579e658e3eee6aac664800cc03f6f7