Pane: use size rather than implicit size to determine the size of the first child
A Pane/Frame is used to frame a control (or a set of controls). Unless
the application sets a custom size on the contentItem, we have
up till now used the implicit size of the first child to also be
the implicit size of the contentItem.

It's hard to see why using the implicit size of a child makes sense
for this calculation. We want the Pane/Frame to wrap the child, but
if the size of the child differs from it's implicit size, this will
simply fail. E.g consider:

 Frame {
    TextArea {
       width: 100
       height: 100
    }
 }

In this case, the frame will be drawn around the implicit size of the
TextArea, and not the TextArea itself.

This patch will change this logic, so that we use the width/height
of the child to calculate the implicit size of the contentItem
instead.

[ChangeLog][QtQuickControls] If the size of a controls
contentItem is calculated from the first child, the childs
size, rather than its implicit size, is now used.

Change-Id: Iea701282a8ed9d581f67c0e055b4ca1d8e846d27